+++ CHANGES.md
@@ -1,5 +1,6 @@
 # 2.0.1 - (XX.XX.XX)
 - fix: map not loading (#289)
+- fix: clean stale years from year filter (#291)
 
 # 2.0.0 - (12.07.26)
 - BREAKING CHANGE: SportTracker is no longer compatible with "georender". https://github.com/deadlocker8/GpxToImageRenderer is used insteaf. See README.md section "Enable GPX preview Images" for detailed information.